examples/perception/vision_language/pixtral_12b/requirements.txt
torch
unsloth
transformers==4.56.2
pillow>=10.0.0
requests>=2.31.0
